[Misra] Modify Misra example
- modify : 10AD/10AE/10AF
- add : 110A

+++ misra/10AF_MIDENTS3/10AF_MIDENTS3__BadCode.c
@@ -1,32 +1,32 @@
-// #include <arm64_neon.h> Ãß°¡
-// main() Ãß°¡
-// 10AF_MIDENTS3_extern.c Ãß°¡
+// #include <arm64_neon.h> ï¿½ß°ï¿½
+// main() ï¿½ß°ï¿½
+// 10AF_MIDENTS3_extern.c ï¿½ß°ï¿½
 
 #include <arm64_neon.h>
 
 void fn1(void)
 {
-	int16_t i; /* Declare an object ""i"" */
+	int i; /* Declare an object ""i"" */
 	{
-		int16_t i; /* Non-compliant - hides previous ""i"" */
+		int i; /* Non-compliant - hides previous ""i"" */
 		i = 3; /* Could be confusing as to which ""i"" this refers */
 	}
 }
 struct astruct
 {
-	int16_t m;
+	int m;
 };
 extern void g(struct astruct* p);
-int16_t xyz = 0; /* Declare an object ""xyz""*/
+int xyz = 0; /* Declare an object ""xyz""*/
 void fn2(struct astruct xyz) /* Non-compliant - outer ""xyz"" is
 * now hidden by parameter name */
 {
 	g(&xyz);
 }
-uint16_t speed;
+int speed;
 void fn3(void)
 {
-	typedef float32_t speed; /* Non-compliant - type hides object */
+	typedef float speed; /* Non-compliant - type hides object */
 }
 int main()
 {
